Order game source files from top-level to innermost
I keep facing issues requiring me to do stuff like
```TypeScript
export let someConst: long.type
defer(() => {
    someConst = some.dependency.makeThing()
}
```
when there wouldn't be any issue with just making it
a legit const if the source files were ordered logically.

- I am yet unwilling to introduce modules/imports/references.
- I don't want to reintroduce some home-brewed dependency
resolution (a la the `dependsOn()` pseudo-method of the past).
- Actually solving the problem by ordering as a human could
is a really difficult problem that I don't want to solve or fake-solve.
- I don't want additional overhead of specifying order manually.
- I don't want some weird naming scheme or other odd paradigm for order.

So this solution is a happy medium.
It gives me the power to structure the game source in an arbitrary,
logical, low-overhead way that still provides some guarantees for order.