Fix bug with horizontal movement always bumping up
Because ground traces previously caused the blocked flag
to propagate all the way up, a body moving horizontally
on top of ground would always bump up before moving over.
This was not noticeable in most cases, but it caused bodies
not to be able to move horizontally when something else was
on top (e.g. stacked boxes).