Fix bug with undefined rendering order
I ran into a scenario with two stacked boxes and the player.
In some cases, the top box would render behind the lower box.

As I understand it, this would occur when the player
was close enough for consideration of rendering order
for both boxes and it would so happen that
the player would be considered behind the lower box
but in front of the top box.
If the player came first in the list part of the walk
before the top box, then the top box would
(incorrectly) render behind the other two sprites.
If the player came after the top box,
then the rendering order would turn out correct.

Since there was a somewhat non-deterministic
rendering order in the case of overlap,
I have changed the algorithm not to pick
an order for overlapping bodies.