Add body dimensions fall-caching
This additional constraint is one of those things
I'm likely to run into later causing a problem and then think to myself,
"Oh yeah, I knew that was going to be a problem,"
except having spent a while first debugging the issue.

If a body's dimensions change, it invalidates the fall cache.