WIP Allow self-level pointer + smooth camera
- Allow pointer trace to reference a the source level
for purposes of e.g. room loops
- Smooth out camera movement between levels so that
there is no jerky frame transition that catches the player's
attention when the level changes

Because I want some levels where the player shouldn't realize
the level has changed, the camera needs to be totally fluid
from one level to the next when the region doesn't change.